syskit: add complete support to get models and transformations from a gazebo instance
This adds a more complete support for SDF and Gazebo in syskit.

Namely, profiles now have a use_sdf_model statement which (1) defines
device(s) for the SDF world and (2) sets the transformer up
accordingly. The transformer is only set for the vehicle's model
while the devices are set for each model in the world (if a world
is selected). This allows to run syskit test without having selected
a full world, but be able to use all of it later.

Worlds and models are resolved using the same rules than the command-line
tools, e.g. 'flat_fish' can resolve 'scenes/flat_fish/flat_fish.world'