syskit: expose more SDF information on the Syskit profile and robot models
This commit cleans up the semantic of 'set of models' (e.g. world)
and 'main robot model', as well as exposes more information out of
the SDF file:

The RobotDefinition API does not deal with worlds anymore. It deals
with a 'main' robot model (the model of the robot the syskit instance
deals with) and the 'exposed' models (the set of models present in
the world). We export everything from the former (sensors, plugins and
links) but expose only the model itself from the latter. The rationale
is that the models are needed for e.g. world visualization but we should
try to access only the information from our system to build the "real"
component network.