syskit: refactor between sdf and gazebo usage
The "*_sdf_*" methods now allow to use SDF models outside gazebo,
i.e. assume that the SDF information represents known information
about the vehicle's world.

The "_*gazebo_*" methods on the other hand assume that the goal is
to interface with a gazebo instance, and define the relevant information
(such as e.g. devices)